+++ .changeset/famous-buttons-hunt.md
@@ -1,0 +1,5 @@
+---
+"@enalmada/clone-code": patch
+---
+
+fix build cli output
